Split code to different files
