--- pet-store-vanilla-js/formElements.js
@@ -1,14 +1,0 @@
-const formElements = {
-    petName: document.getElementById('petName'),
-    age: document.getElementById('age'),
-    notes: document.getElementById('notes'),
-    kind: document.getElementById('kind'),
-    healthProblems: document.getElementById('healthProblems'),
-    addedDatePicker: document.getElementById('addedDatePicker'),
-    addedDateText: document.getElementById('addedDateText'),
-    saveButton: document.getElementById('form-save-btn'),
-    deleteButton: document.getElementById('form-delete-btn'),
-    cancelButton: document.getElementById('form-cancel-btn'),
-};
-
-export default formElements;
